Move to a lerna project
