+++ .eslintignore
@@ -1,0 +1,2 @@
+packages/**/dist
+packages/**/node_modules
