+++ .gitignore
@@ -1,24 +1,18 @@
 # See https://help.github.com/ignore-files/ for more about ignoring files.
 
 # compiled output
-/dist
-/output
-/tmp
+/packages/**/dist
+/packages/**/tmp
 
 # dependencies
 /node_modules
-/bower_components
+/packages/**/node_modules
 
 # misc
-/.sass-cache
 /connect.lock
 /coverage/*
 /libpeerconnection.log
 npm-debug.log*
 yarn-error.log
+lerna-debug.log
 testem.log
-
-# ember-try
-.node_modules.ember-try/
-bower.json.ember-try
-package.json.ember-try
