+++ package.json
@@ -1,69 +1,23 @@
 {
   "name": "testdouble-qunit",
   "version": "1.0.1",
-  "description": "Verify testdouble stubs with QUnit",
-  "keywords": [
-    "ember-addon"
-  ],
-  "license": "MIT",
-  "author": "",
-  "directories": {
-    "doc": "doc",
-    "test": "tests"
-  },
   "homepage": "https://alexlafroscia.com/testdouble-qunit",
   "repository": "https://github.com/alexlafroscia/testdouble-qunit",
+  "private": true,
+  "workspaces": [
+    "packages/*"
+  ],
   "scripts": {
-    "build": "rollup -c",
-    "lint": "eslint ./*.js addon-test-support config lib server test-support tests",
-    "prepublish": "rollup -c",
-    "start": "ember serve",
-    "test": "ember test"
-  },
-  "dependencies": {
-    "broccoli-funnel": "^2.0.1",
-    "broccoli-merge-trees": "^3.0.0",
-    "ember-cli-babel": "^6.6.0"
+    "deploy": "packages/ember-cli-testdouble-qunit/node_modules/.bin/ember deploy production",
+    "lint": "eslint .",
+    "test": "lerna run test"
   },
   "devDependencies": {
-    "broccoli-asset-rev": "^2.4.5",
-    "ember-ajax": "^3.0.0",
-    "ember-cli": "~3.0.2",
-    "ember-cli-addon-docs": "^0.3.0",
-    "ember-cli-addon-docs-yuidoc": "^0.1.1",
-    "ember-cli-dependency-checker": "^2.0.0",
-    "ember-cli-deploy": "^1.0.2",
-    "ember-cli-deploy-build": "^1.1.1",
-    "ember-cli-deploy-git": "^1.3.2",
-    "ember-cli-deploy-git-ci": "^1.0.1",
-    "ember-cli-htmlbars": "^2.0.1",
-    "ember-cli-htmlbars-inline-precompile": "^1.0.0",
-    "ember-cli-inject-live-reload": "^1.4.1",
-    "ember-cli-qunit": "^4.1.1",
-    "ember-cli-shims": "^1.2.0",
-    "ember-cli-sri": "^2.1.0",
-    "ember-cli-testdouble": "^0.1.3",
-    "ember-cli-uglify": "^2.0.0",
-    "ember-disable-prototype-extensions": "^1.1.2",
-    "ember-export-application-global": "^2.0.0",
-    "ember-load-initializers": "^1.0.0",
-    "ember-maybe-import-regenerator": "^0.1.6",
-    "ember-resolver": "^4.0.0",
-    "ember-source": "~3.0.0",
     "eslint": "^4.19.1",
     "eslint-config-prettier": "^2.9.0",
-    "eslint-plugin-ember": "^5.0.0",
     "eslint-plugin-node": "^5.2.1",
     "eslint-plugin-prettier": "^2.6.0",
-    "loader.js": "^4.2.3",
-    "prettier": "^1.11.1",
-    "rollup": "^0.57.1",
-    "rollup-plugin-babel": "^3.0.3"
-  },
-  "engines": {
-    "node": "6.* || >= 7.*"
-  },
-  "ember-addon": {
-    "configPath": "tests/dummy/config"
+    "lerna": "^2.9.0",
+    "prettier": "^1.11.1"
   }
 }
