+++ packages/ember-cli-testdouble-qunit/.watchmanconfig
