+++ packages/ember-cli-testdouble-qunit/config/addon-docs.js
