+++ packages/ember-cli-testdouble-qunit/index.js
@@ -1,0 +1,14 @@
+'use strict';
+
+module.exports = {
+  name: 'testdouble-qunit',
+
+  included() {
+    this._super.included.apply(this, arguments);
+
+    this.import('node_modules/testdouble-qunit/dist/testdouble-qunit.umd.js', {
+      type: 'test',
+      using: [{ transformation: 'amd', as: 'testdouble-qunit' }]
+    });
+  }
+};
