+++ packages/ember-cli-testdouble-qunit/package.json
@@ -1,0 +1,61 @@
+{
+  "name": "ember-cli-testdouble-qunit",
+  "version": "1.0.1",
+  "description": "Verify testdouble stubs with QUnit",
+  "keywords": [
+    "ember-addon",
+    "testdouble",
+    "qunit"
+  ],
+  "license": "MIT",
+  "author": "Alex LaFroscia <alex@lafroscia.com.",
+  "directories": {
+    "doc": "doc",
+    "test": "tests"
+  },
+  "homepage": "https://alexlafroscia.com/testdouble-qunit",
+  "repository": "https://github.com/alexlafroscia/testdouble-qunit",
+  "scripts": {
+    "start": "ember serve",
+    "test": "ember test"
+  },
+  "dependencies": {
+    "ember-cli-babel": "^6.6.0",
+    "testdouble-qunit": "1.0.1"
+  },
+  "devDependencies": {
+    "broccoli-asset-rev": "^2.4.5",
+    "ember-ajax": "^3.0.0",
+    "ember-cli": "~3.0.2",
+    "ember-cli-addon-docs": "^0.3.0",
+    "ember-cli-addon-docs-yuidoc": "^0.1.1",
+    "ember-cli-dependency-checker": "^2.0.0",
+    "ember-cli-deploy": "^1.0.2",
+    "ember-cli-deploy-build": "^1.1.1",
+    "ember-cli-deploy-git": "^1.3.2",
+    "ember-cli-deploy-git-ci": "^1.0.1",
+    "ember-cli-htmlbars": "^2.0.1",
+    "ember-cli-htmlbars-inline-precompile": "^1.0.0",
+    "ember-cli-inject-live-reload": "^1.4.1",
+    "ember-cli-qunit": "^4.1.1",
+    "ember-cli-shims": "^1.2.0",
+    "ember-cli-sri": "^2.1.0",
+    "ember-cli-testdouble": "^0.1.3",
+    "ember-cli-uglify": "^2.0.0",
+    "ember-disable-prototype-extensions": "^1.1.2",
+    "ember-export-application-global": "^2.0.0",
+    "ember-load-initializers": "^1.0.0",
+    "ember-maybe-import-regenerator": "^0.1.6",
+    "ember-resolver": "^4.0.0",
+    "ember-source": "~3.0.0",
+    "ember-source-channel-url": "^1.0.1",
+    "ember-try": "^0.2.23",
+    "loader.js": "^4.2.3"
+  },
+  "engines": {
+    "node": "6.* || >= 7.*"
+  },
+  "ember-addon": {
+    "configPath": "tests/dummy/config"
+  }
+}
