+++ packages/ember-cli-testdouble-qunit/tests/dummy/app/app.js
