+++ packages/ember-cli-testdouble-qunit/tests/dummy/app/docs/index/template.md
