+++ packages/ember-cli-testdouble-qunit/tests/dummy/app/docs/installation/template.md
