+++ packages/ember-cli-testdouble-qunit/tests/dummy/app/docs/template.hbs
