+++ packages/ember-cli-testdouble-qunit/tests/dummy/app/resolver.js
