+++ packages/ember-cli-testdouble-qunit/tests/dummy/config/targets.js
